[refacto][server] Clean req.session
restore req.session.ap

+++ packages/infoportal-server/src/server/controller/ControllerWorkspace.ts
@@ -1,6 +1,8 @@
 import {NextFunction, Request, Response} from 'express'
 import {PrismaClient} from '@prisma/client'
 import {WorkspaceService} from '../../feature/workspace/WorkspaceService.js'
+import {isAuthenticated} from '../Routes.js'
+import {AppError} from '../../helper/Errors.js'
 
 export class ControllerWorkspace {
   constructor(
@@ -8,14 +10,26 @@ export class ControllerWorkspace {
     private service = new WorkspaceService(prisma),
   ) {}
 
+  readonly checkSlug = async (req: Request, res: Response, next: NextFunction) => {
+    if (!isAuthenticated(req)) throw new AppError.Forbidden()
+    const {slug} = await WorkspaceService.schema.slug.validate(req.body)
+    const suggestedSlug = await this.service.getUniqSlug(slug)
+    res.send({
+      isFree: slug === suggestedSlug,
+      suggestedSlug,
+    })
+  }
+
   readonly create = async (req: Request, res: Response, next: NextFunction) => {
+    if (!isAuthenticated(req)) throw new AppError.Forbidden()
     const body = await WorkspaceService.schema.create.validate(req.body)
-    const email = req.session.session?.email ?? 'unknown'
-    const data = await this.service.create(body, email)
+    const connectedUser =req.session.app.user
+    const data = await this.service.create(body, connectedUser)
     res.send(data)
   }
 
   readonly update = async (req: Request, res: Response, next: NextFunction) => {
+    if (!isAuthenticated(req)) throw new AppError.Forbidden()
     const {id} = await WorkspaceService.schema.id.validate(req.params)
     const body = await WorkspaceService.schema.update.validate(req.body)
     const data = await this.service.update(id, body)
@@ -23,6 +37,7 @@ export class ControllerWorkspace {
   }
 
   readonly remove = async (req: Request, res: Response, next: NextFunction) => {
+    if (!isAuthenticated(req)) throw new AppError.Forbidden()
     const {id} = await WorkspaceService.schema.id.validate(req.params)
     await this.service.remove(id)
     res.send()
